Fixing the dashboard use case.
