+++ src/main/java/com/securian/creditcompass/Dashboard/DashboardInputBoundary.java
@@ -1,4 +1,7 @@
 package com.securian.creditcompass.dashboard;
 
-public class DashboardInputBoundary {
+import java.util.List;
+
+public interface DashboardInputBoundary {
+    List<List<Object>> execute(DashboardInputData dashboardInputData);
 }
