+++ src/main/java/com/securian/creditcompass/Dashboard/DashboardInputData.java
@@ -1,4 +1,13 @@
 package com.securian.creditcompass.dashboard;
 
+import lombok.Getter;
+
+@Getter
 public class DashboardInputData {
+    final private String username;
+
+    public DashboardInputData(String username) {
+        this.username = username;
+    }
+
 }
