+++ src/main/java/com/securian/creditcompass/Dashboard/DashboardInteractor.java
@@ -1,49 +1,43 @@
 package com.securian.creditcompass.Dashboard;
 
+import com.securian.creditcompass.OrderCalculator.OrderCalculator;
+import com.securian.creditcompass.dashboard.DashboardInputBoundary;
+import com.securian.creditcompass.dashboard.DashboardInputData;
+import com.securian.creditcompass.dataAccess.ClaimRepository;
 import com.securian.creditcompass.entities.Claim;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
-public class DashboardInteractor {
+public class DashboardInteractor implements DashboardInputBoundary {
+    private final ClaimRepository claimRepository;
 
-    public List<List<Object>> findAttributes(List<Claim> claimList) {
-        List<List<Object>> allClaimsAttributes = new ArrayList<>();
+    public DashboardInteractor(ClaimRepository claimRepository) {
+        this.claimRepository = claimRepository;
+    }
+
+    @Override
+    public List<List<Object>> execute(DashboardInputData dashboardInputData) {
+
+        // Find all the claims associated with the given examiner. Notice that the examiner might not
+        // have any claims associated with them.
+        Optional<List<Claim>> claimsList = claimRepository.findClaimsByExaminer(dashboardInputData.getUsername());
+
+        //TODO: Do we need to change OrderCalculator to a helper method? Furthermore, OrderCalculator
+        // does not need to be an Iterator.
 
 
-        for (Claim claim : claimList) {
-            List<Object> claimAttributes = new ArrayList<>();
-            // Assuming Claim class has attributes like claimId, amount, date, etc.
-            claimAttributes.add(claim.getId());
-            claimAttributes.add(claim.getClaimAmt());
-            claimAttributes.add(claim.getCreationDateTime());
-            claimAttributes.add(claim.getClaimType());
-            double urgency = claim.getUrgencyScore();
-            if (urgency >= 6) {
-                claimAttributes.add("High");
-            }
-            else if (urgency >= 3) {
-                claimAttributes.add("Medium");
-            }
-            else {
-                claimAttributes.add("Low");
-            }
-            double complexity = claim.getComplexityScore();
-            // Complexity
-            if (complexity >= 7) {
-                claimAttributes.add("High");
-            }
-            else if (complexity >= 4) {
-                claimAttributes.add("Medium");
-            }
-            else {
-                claimAttributes.add("Low");
-            }
-            allClaimsAttributes.add(claimAttributes);
+        // If the examiner has claims:
+        if (claimsList.isPresent()) {
+            //
+            OrderCalculator scoredClaims = new OrderCalculator(claims.get());
+            List<Claim> sortedClaims = scoredClaims.getOrderedClaims();
+            return renderDashboardService.findAttributes(sortedClaims);
+        } else {
+            return new ArrayList<>();
         }
-
-        return allClaimsAttributes;
     }
 }
