--- src/main/java/com/securian/creditcompass/Dashboard/DashboardOutputBoundary.java
@@ -1,4 +1,0 @@
-package com.securian.creditcompass.dashboard;
-
-public class DashboardOutputBoundary {
-}
