+++ src/main/java/com/securian/creditcompass/dataAccess/ClaimRepository.java
@@ -5,12 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Repository
 public interface ClaimRepository extends JpaRepository<Claim, Long> {
 
-    List<Claim> findClaimsByExaminer(String examiner);
+    Optional<List<Claim>> findClaimsByExaminer(String examiner);
     List<Claim> findAll();
 
 }
